+++ src/asyncworker.h
@@ -7,13 +7,13 @@ namespace MurmurHash {
   using v8::String;
   using v8::Uint32;
 
-  template<MurmurHashFunctionType HashFunction, typename HashValueType, int32_t HashLength>
+  template<MurmurHashFunctionType HashFunction, typename HashValueType, int32_t HashLength, ByteOrderType OutputByteOrder>
   class MurmurHashWorker : public Nan::AsyncWorker {
     public:
         NAN_INLINE MurmurHashWorker(Nan::Callback *callback);
         NAN_INLINE MurmurHashWorker(Nan::Callback *callback, OutputType outputType, uint32_t seed,
                                       Local<Value> key, const enum Nan::Encoding encoding, const bool validEncoding);
-        void SaveOutputBuffer(const Local<Value> &buffer, int32_t offset, int32_t length);
+        NAN_INLINE void SaveOutputBuffer(const Local<Value> &buffer, int32_t offset, int32_t length);
         void Execute();
         void HandleOKCallback();
         void HandleErrorCallback();
@@ -25,6 +25,7 @@ namespace MurmurHash {
         int32_t offset_;
         int32_t length_;
         HashValueType hash_[HashLength];
+        char dataBuffer[NODE_MURMURHASH_KEY_BUFFER_SIZE];
   };
 
 }