+++ src/incremental/hasher.cc
@@ -1,7 +1,7 @@
 #include "static_assert.h"
 #include "hasher.h"
-#include "inputdata.h"
 #include "murmurhashutils.h"
+#include "inputdata.h"
 #include "hasher_impl.h"
 #include "asyncupdate.h"
 
@@ -28,11 +28,12 @@ namespace MurmurHash {
    *
    * Create MurmurHash utility
    * 
-   * new MurmurHash([seed|hash|serial])
+   * new MurmurHash([seed|hash|serial][, endianness="BE"])
    *
    * @param {number} seed - initial murmur hash seed as 32 bit integer
    * @param {MurmurHash} hash - an instance of another MurmurHash of the same type
    * @param {string|Buffer} serial - serialized state of the same MurmurHash type
+   * @param {string} endianness - digest byte order: "BE", "LE" or "platform"
   **/
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
   NAN_METHOD(SINGLE_ARG(IncrementalHasher<H,HashValueType,HashLength>::New))
@@ -40,9 +41,11 @@ namespace MurmurHash {
 
     IncrementalHasher_T *self;
 
+    const int argc = std::min(2, info.Length());
+
     if ( info.IsConstructCall() ) {
 
-      if ( info.Length() > 0 ) {
+      if ( argc > 0 ) {
         if ( info[0]->IsNumber() ) {
           // seed
           uint32_t seed = Nan::To<uint32_t>(info[0]).FromMaybe(0U);
@@ -89,6 +92,13 @@ namespace MurmurHash {
           return Nan::ThrowTypeError("Expected a seed number, MurmurHash instance or serialized state");
         }
 
+        if ( argc > 1 && !info[1]->IsUndefined() && !info[1]->IsNull() ) {
+          if ( ! self->SetEndiannessFrom( info[1] ) ) {
+            delete self;
+            return;
+          }
+        }
+
       } else {
         self = new IncrementalHasher_T();
       }
@@ -97,9 +107,8 @@ namespace MurmurHash {
       info.GetReturnValue().Set( info.This() );
 
     } else {
-      int argc = std::min(1, info.Length());
-      Local<Value> argv[1];
-      if (argc > 0) argv[0] = info[0];
+      Local<Value> argv[2];
+      for (int i = argc; i-- > 0; argv[i] = info[i]);
       Local<Function> cons = Nan::GetFunction(Nan::New(constructor)).ToLocalChecked();
       MaybeLocal<Object> mayinst = Nan::NewInstance(cons, argc, &argv[0]);
       if ( ! mayinst.IsEmpty() ) {
@@ -109,9 +118,11 @@ namespace MurmurHash {
   }
 
   /**
-   * Copy the internal state onto the target utility instance
+   * Copy the internal state onto the target utility instance.
    * 
    * copy(target)
+   *
+   * This method does not alter target endianness.
    * 
    * @param {MurmurHash} target - a different instance of MurmurHash utility
    *                              of the same type
@@ -173,7 +184,7 @@ namespace MurmurHash {
 
     OutputType outputType( DefaultOutputType );
 
-    int argc = info.Length();
+    const int argc = info.Length();
 
     if ( argc > 0 ) {
       if ( info[0]->IsString() ) {
@@ -190,43 +201,10 @@ namespace MurmurHash {
 
     self->Digest( hash );
 
-    switch(outputType) {
-      case HexStringOutputType:
-        result = HashToHexString<HashLength>( hash );
-        break;
-
-      case BinaryStringOutputType:
-        result = HashToEncodedString<HashLength>( hash, Nan::BINARY );
-        break;
-
-      case Base64StringOutputType:
-        result = HashToEncodedString<HashLength>( hash, Nan::BASE64 );
-        break;
-
-      case NumberOutputType:
-        if (HashSize == sizeof(uint32_t)) {
-          result = Nan::New<Uint32>( (uint32_t) (*hash) );
-        } else {
-          result = HashToHexString<HashLength>( hash );
-        }
-        break;
-
-      case ProvidedBufferOutputType:
-        result = info[0];
-        WriteHashToBuffer<HashLength>(
-              hash,
-              node::Buffer::Data(result),
-              (int32_t) node::Buffer::Length(result),
-              (argc > 1) ? Nan::To<int32_t>(info[1]).FromMaybe(0) : 0,
-              (argc > 2)
-              ? Nan::To<int32_t>(info[2]).FromMaybe(HashSize)
-              : HashSize);
-        break;
-
-      default:
-        result = Nan::NewBuffer( HashSize ).ToLocalChecked();
-        WriteHashBytes<HashLength>(hash, (uint8_t *) node::Buffer::Data(result));
-        break;
+    if (self->outputByteOrder == MSBFirst) {
+      Output<MSBFirst>(hash, outputType, argc, info, result);
+    } else {
+      Output<LSBFirst>(hash, outputType, argc, info, result);
     }
 
     info.GetReturnValue().Set( result );
@@ -298,8 +276,6 @@ namespace MurmurHash {
    * @param {Function} callback - optional callback(err)
    *       if provided the hash will be updated asynchronously using libuv
    *       worker queue, the return value in this instance will be `undefined`
-   *       Be carefull as reading and writing by multiple threads to the same
-   *       memory may render undetermined results
    * @return {MurmurHash} this
   **/
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
@@ -370,6 +346,25 @@ namespace MurmurHash {
   }
 
   /**
+   * @property {string} endianness - digest byte order: "BE", "LE" or "platform"
+  **/
+  template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
+  NAN_GETTER(SINGLE_ARG(IncrementalHasher<H,HashValueType,HashLength>::GetEndianness))
+  {
+    IncrementalHasher_T *self = ObjectWrap::Unwrap<IncrementalHasher_T>( info.Holder() );
+    info.GetReturnValue().Set(
+      Nan::New<String>(self->outputByteOrder == MSBFirst ? "BE" : "LE").ToLocalChecked()
+    );
+  }
+
+  template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
+  NAN_SETTER(SINGLE_ARG(IncrementalHasher<H,HashValueType,HashLength>::SetEndianness))
+  {
+    IncrementalHasher_T *self = ObjectWrap::Unwrap<IncrementalHasher_T>( info.Holder() );
+    (void) self->SetEndiannessFrom( value );
+  }
+
+  /**
    * @property {boolean} isBusy - is asynchronous update in progress
   **/
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
@@ -400,20 +395,20 @@ namespace MurmurHash {
 
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
   NAN_INLINE IncrementalHasher<H,HashValueType,HashLength>
-  ::IncrementalHasher(const uint32_t seed) : hasher(seed), total(0), asyncInProgress(false) {};
+  ::IncrementalHasher(const uint32_t seed) :
+    hasher(seed), total(0), outputByteOrder(MSBFirst), asyncInProgress(false) {};
 
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
   NAN_INLINE IncrementalHasher<H,HashValueType,HashLength>
   ::IncrementalHasher(const IncrementalHasher_T& other) : ObjectWrap(),
-                                                          hasher(other.hasher),
-                                                          total(other.total),
-                                                          asyncInProgress(false) {};
+    hasher(other.hasher), total(other.total), outputByteOrder(other.outputByteOrder), asyncInProgress(false) {};
 
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
   NAN_INLINE IncrementalHasher<H,HashValueType,HashLength>
-  ::IncrementalHasher(const uint8_t *serial) : hasher(serial), asyncInProgress(false)
+  ::IncrementalHasher(const uint8_t *serial) :
+    hasher(serial), outputByteOrder(MSBFirst), asyncInProgress(false)
   {
-    ReadHashBytes<1>(&serial[kHashSerialTotalIndex], &total);
+    ReadHashBytesMSB<1>(&serial[kHashSerialTotalIndex], &total);
   }
 
   /*--------------- static methods ----------------*/
@@ -442,6 +437,52 @@ namespace MurmurHash {
     return false;
   }
 
+  template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
+  template<ByteOrderType OutputByteOrder>
+  NAN_INLINE void IncrementalHasher<H,HashValueType,HashLength>
+  ::Output(const HashValueType hash[HashLength], const OutputType &outputType,
+           const int &argc, const Nan::NAN_METHOD_ARGS_TYPE info, Local<Value> &result)
+  {
+    switch(outputType) {
+      case HexStringOutputType:
+        result = HashToEncodedString<OutputByteOrder, HashLength>( hash, Nan::HEX );
+        break;
+
+      case BinaryStringOutputType:
+        result = HashToEncodedString<OutputByteOrder, HashLength>( hash, Nan::BINARY );
+        break;
+
+      case Base64StringOutputType:
+        result = HashToEncodedString<OutputByteOrder, HashLength>( hash, Nan::BASE64 );
+        break;
+
+      case NumberOutputType:
+        if (HashSize == sizeof(uint32_t)) {
+          result = Nan::New<Uint32>( (uint32_t) (*hash) );
+        } else {
+          result = HashToEncodedString<OutputByteOrder, HashLength>( hash, Nan::HEX );
+        }
+        break;
+
+      case ProvidedBufferOutputType:
+        result = info[0];
+        WriteHashToBuffer<OutputByteOrder, HashLength>(
+              hash,
+              node::Buffer::Data(result),
+              (int32_t) node::Buffer::Length(result),
+              (argc > 1) ? Nan::To<int32_t>(info[1]).FromMaybe(0) : 0,
+              (argc > 2)
+              ? Nan::To<int32_t>(info[2]).FromMaybe(HashSize)
+              : HashSize);
+        break;
+
+      default:
+        result = Nan::NewBuffer( HashSize ).ToLocalChecked();
+        WriteHashBytes<OutputByteOrder, HashLength>(hash, (uint8_t *) node::Buffer::Data(result));
+        break;
+    }
+  }
+
   /*-------------- instance methods ---------------*/
 
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
@@ -464,7 +505,7 @@ namespace MurmurHash {
 
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
   NAN_INLINE bool IncrementalHasher<H,HashValueType,HashLength>
-  ::CheckAsyncUpdateInProgress()
+  ::CheckAsyncUpdateInProgress() const
   {
     if (asyncInProgress) {
       Nan::ThrowError(MESSAGE_ERROR_PENDING_UPDATE);
@@ -489,7 +530,7 @@ namespace MurmurHash {
     // write state type
     serial[kHashSerialTypeIndex] |= kHashSerialType;
     // write total
-    WriteHashBytes<1>(&total, &serial[kHashSerialTotalIndex]);
+    WriteHashBytesMSB<1>(&total, &serial[kHashSerialTotalIndex]);
     // build checksum
     checksum_t chksum = PMurHash32(serial, kHashSerialSize - kHashSerialCkSize, kHashSerialCkSeed);
     if (kHashSerialCkSize < sizeof(checksum_t)) {
@@ -500,6 +541,27 @@ namespace MurmurHash {
       serial[--i] = (uint8_t) chksum & 0xFF;
       if (i == kHashSerialCkIndex) break;
     }
+  }
+
+  template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
+  bool IncrementalHasher<H,HashValueType,HashLength>
+  ::SetEndiannessFrom(const Local<Value> &value)
+  {
+    Nan::Utf8String strval( value );
+    if ( *strval != NULL ) {
+      if ( StringEqualLower(*strval, "be") ) {
+        outputByteOrder = MSBFirst;
+        return true;
+      } else if ( StringEqualLower(*strval, "le") ) {
+        outputByteOrder = LSBFirst;
+        return true;
+      } else if ( StringEqualLower(*strval, "platform") ) {
+        outputByteOrder = IsBigEndian() ? MSBFirst : LSBFirst;
+        return true;
+      }
+    }
+    Nan::ThrowTypeError("Unknown endianness: should be \"BE\", \"LE\" or \"platform\"");
+    return false;
   }
 
   template<template <typename,int32_t>class H, typename HashValueType, int32_t HashLength>
@@ -537,10 +599,15 @@ namespace MurmurHash {
     Local<ObjectTemplate> i_t = tpl->InstanceTemplate();
     i_t->SetInternalFieldCount(1);
 
-    Nan::SetAccessor( i_t, Nan::New<String>("isBusy").ToLocalChecked(), GetIsBusy,
-                            NULL, Local<Value>(), v8::DEFAULT,
-                            static_cast<PropertyAttribute>(DontEnum | DontDelete));
-    Nan::SetAccessor( i_t, Nan::New<String>("total").ToLocalChecked(), GetTotal );
+    Nan::SetAccessor( i_t, Nan::New<String>("isBusy").ToLocalChecked(),
+                          GetIsBusy, NULL, Local<Value>(), v8::DEFAULT,
+                          static_cast<PropertyAttribute>(DontEnum | DontDelete));
+    Nan::SetAccessor( i_t, Nan::New<String>("total").ToLocalChecked(),
+                          GetTotal, NULL, Local<Value>(), v8::DEFAULT,
+                          static_cast<PropertyAttribute>(DontDelete));
+    Nan::SetAccessor( i_t, Nan::New<String>("endianness").ToLocalChecked(),
+                          GetEndianness, SetEndianness, Local<Value>(), v8::DEFAULT,
+                          static_cast<PropertyAttribute>(DontDelete));
    
     Nan::SetTemplate(tpl, Nan::New<String>("SERIAL_BYTE_LENGTH").ToLocalChecked(),
                           Nan::New<Int32>(kHashSerialSize),
@@ -574,6 +641,7 @@ namespace MurmurHash {
     IncrementalHasher<IncrementalMurmurHash128, uint32_t, 4>::Init(target, "MurmurHash128x86");
   #endif
   }
+
 }
 
 NODE_MODULE(murmurhashincremental, MurmurHash::Init)