+++ src/incremental/hasher.h
@@ -4,6 +4,8 @@
 #include "nodemurmurhash.h"
 
 namespace MurmurHash {
+  using v8::Local;
+  using v8::Value;
   using v8::FunctionTemplate;
   using Nan::Persistent;
   using Nan::ObjectWrap;
@@ -18,7 +20,7 @@ namespace MurmurHash {
       typedef uint32_t checksum_t;
 
       static void Init(Nan::ADDON_REGISTER_FUNCTION_ARGS_TYPE& target,
-                                 const char* name, const char *altname = NULL);
+                                    const char* name, const char *altname = NULL);
 
       static NAN_METHOD(New);
       static NAN_METHOD(Copy);
@@ -26,18 +28,22 @@ namespace MurmurHash {
       static NAN_METHOD(Serialize);
       static NAN_METHOD(Update);
 
+      static NAN_GETTER(GetEndianness);
+      static NAN_SETTER(SetEndianness);
       static NAN_GETTER(GetIsBusy);
       static NAN_GETTER(GetTotal);
 
-      static NAN_INLINE bool IsSerialTypeValid(uint8_t *serial);
+      NAN_INLINE static bool IsSerialTypeValid(uint8_t *serial);
 
       static Persistent<FunctionTemplate> constructor;
 
       NAN_INLINE void AsyncUpdateComplete(void);
-      NAN_INLINE bool CheckAsyncUpdateInProgress(void);
+      NAN_INLINE bool CheckAsyncUpdateInProgress(void) const;
       NAN_INLINE void Digest(HashValueType *hash) const;
       NAN_INLINE void Serialize(uint8_t *serial) const;
       NAN_INLINE void Update(const void *data, uint32_t length);
+
+      char dataBuffer[NODE_MURMURHASH_KEY_BUFFER_SIZE];
 
     private:
       NAN_INLINE IncrementalHasher(const uint32_t seed = 0U);
@@ -45,10 +51,16 @@ namespace MurmurHash {
       NAN_INLINE IncrementalHasher(const IncrementalHasher_T& other);
       NAN_INLINE void operator=(const IncrementalHasher_T&);
 
+      template<ByteOrderType OutputByteOrder>
+      NAN_INLINE static void Output(const HashValueType hash[HashLength], const OutputType &outputType,
+                            const int &argc, const Nan::NAN_METHOD_ARGS_TYPE info, Local<Value> &result);
+
       NAN_INLINE bool AsyncUpdateBegin(void);
+      bool SetEndiannessFrom(const Local<Value> &value);
 
       H<HashValueType,HashLength> hasher;
       total_t total;
+      ByteOrderType outputByteOrder;
       bool asyncInProgress;
 
       #define BASE64_ENCODED_SIZE(size) ((size + 2 - ((size + 2) % 3)) / 3 * 4)